Moved clipboard api to data layer
