+++ PWManager.Data/DependencyInjection.cs
@@ -2,6 +2,7 @@
 using PWManager.Application.Services.Interfaces;
 using PWManager.Data.Repositories;
 using PWManager.Data.Services;
+using PWManager.Data.System;
 using PWManager.Domain.Repositories;
 
 namespace PWManager.Data; 
@@ -16,6 +17,8 @@ public static class DependencyInjection {
 
         services.AddTransient<ILoginService, LoginService>();
         services.AddTransient<IDatabaseInitializerService, DatabaseInitializerService>();
+        
+        services.AddTransient<IClipboard, Clipboard>();
         return services;
     }
 }
