+++ PWManager.Data/System/OperatingSystem.cs
@@ -1,12 +1,12 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace PWManager.CLI.Abstractions; 
+namespace PWManager.Data.System; 
 
 public static class OperatingSystem {
     public static bool IsWindows() =>
         RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
 
-    public static bool IsMacOS() =>
+    public static bool IsMacOs() =>
         RuntimeInformation.IsOSPlatform(OSPlatform.OSX);
 
     public static bool IsLinux() =>
